Cart

+++ Shop/Shop/Common/UserLogin.cs
@@ -9,7 +9,7 @@ namespace Shop.Common
         [Serializable]
         public class UserLogin
         {
-            public long UserID { get; set; }
+            public int UserID { get; set; }
             public string UserName { get; set; }
             public string Role { get; set; }
         }
